+++ crypto-onramp-example/src/main/java/com/stripe/android/crypto/onramp/example/ui/screens/AuthenticatedOperationsScreen.kt
@@ -36,7 +36,7 @@ internal fun AuthenticatedOperationsScreen(
     onCollectKyc: (KycInfo) -> Unit,
     onVerifyKyc: () -> Unit,
     onStartVerification: () -> Unit,
-    onShowCrsCarfDeclaration: () -> Unit,
+    onShowUserAttestation: () -> Unit,
     onCollectPayment: (PaymentMethodSelection) -> Unit,
     onCreatePaymentToken: () -> Unit,
     onCreateSession: () -> Unit,
@@ -149,7 +149,7 @@ internal fun AuthenticatedOperationsScreen(
 
         VerificationSection(
             onStartVerification = onStartVerification,
-            onShowCrsCarfDeclaration = onShowCrsCarfDeclaration
+            onShowUserAttestation = onShowUserAttestation
         )
         PaymentSection(
             googlePayIsReady = uiState.googlePayIsReady,